Changed vol() to return size_t

+++ include/qdp_layout.h
@@ -83,7 +83,7 @@ namespace QDP {
 		const multi1d<int>& lattSize() QDP_CONST;
 
 		//! Total lattice volume
-		int vol() QDP_CONST;
+		size_t  vol() QDP_CONST;
 
 		//! Number of sites on node
 		int sitesOnNode() QDP_CONST;
